+++ packages/atomic/src/components/common/atomic-result/atomic-result.tsx
@@ -4,18 +4,15 @@ import {applyFocusVisiblePolyfill} from '../../../utils/initialization-utils';
 import {
   DisplayConfig,
   ResultContextEvent,
-} from '../result-template-components/result-template-decorators';
-import {ResultRenderingFunction} from '../../common/result-list/result-list-common';
+} from '../../search/result-template-components/result-template-decorators';
+import {ResultRenderingFunction} from '../result-list/result-list-common';
 import {
   ResultLayout,
   ResultDisplayDensity,
   ResultDisplayImageSize,
   ResultDisplayLayout,
-} from '../../common/layout/display-options';
-import {
-  AtomicCommonStore,
-  AtomicCommonStoreData,
-} from '../../common/interface/store';
+} from '../layout/display-options';
+import {AtomicCommonStore, AtomicCommonStoreData} from '../interface/store';
 
 /**
  * The `atomic-result` component is used internally by the `atomic-result-list` component.
